refactor(torghut): split status health payload modules
